Move port handling out of the oauth2 file
